Add a new package archive.bt and move the functionality of the backup-tool command line script into modules in that package
